fix(activity): issues with starting an activity are now possible to handle by users
Previously, when the provider failed to start the activity, this was not
easily visible for the requestor. Now the issues with creating the
activity on the provider are visible to the user and lead to proper
activity and agreement termination. However, if the activity could not
be created because the provider is a faulty one, the agreement is still
not terminated - at least not by the task-executor.