fix(payments): preventing double payments resulting from a race condition
Fixed the situation where invoice and debit notes sent at the end of the
agreement resulted with double payments. While the original issue has to
be solved in yagna, we provide an integrity ensuring solution right
now.

Doubled payments can still occur, if the debit note will arrive
before the invoice. Solving this was not part of this particular change.