fix(yagna): enforce min version of supported yagna to be 0.15.0
The change of the default test network to holesky is a breaking change.
This commit introduces the min supported version of yagna check which
throws an error if the user is running an older than supported version
of yagna.

BREAKING CHANGE: The SDK now requires yagna at least in version 0.15.0